+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+##
+
+clear chartInstance when componentWillReceiveProps in LineChart
+
+## 1.1.4
+
+hotfix bug in 1.1.3
+
+## 1.1.3
+
+clear instance when componentWillReceiveProps in KLineChart
+
 ## 1.1.2
 
 Add xLabelCallback in KLineChart